requirements.txt
# ---- PyTorch (GPU, CUDA 12.4) ----
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124
torchaudio==2.6.0+cu124

# ---- HuggingFace / OCR ----
transformers==4.57.3
accelerate==1.12.0
datasets==4.4.1
evaluate==0.4.6
sentencepiece==0.2.1
tokenizers==0.22.1
huggingface-hub==0.36.0
jiwer==4.0.0

# ---- Data / Utilities ----
numpy==2.1.2
pandas==2.3.3
tqdm==4.67.1
python-dateutil==2.9.0.post0
pytz==2025.2

# ---- Image ----
pillow==11.3.0

# ---- Visualization ----
matplotlib==3.10.7
tensorboard==2.20.0

# ---- OPTIONAL: semantic similarity (used in some metrics)
sentence-transformers==5.1.2
scikit-learn==1.7.2

# ---- Required low-level deps ----
protobuf==6.33.1
pyarrow==22.0.0
joblib==1.5.2
filelock==3.19.1
requests==2.32.5
packaging==25.0
